+++ src/test/java/domain/CardDeckGeneratorTest.java
@@ -7,13 +7,13 @@ import java.util.Queue;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class CardGeneratorTest {
+public class CardDeckGeneratorTest {
 
     @Test
     @DisplayName("52개의 카드를 생성한다.")
     void createCard() {
-        CardGenerator cardGenerator = new CardGenerator();
-        Queue<Card> cards = cardGenerator.generate();
+        CardDeckGenerator cardDeckGenerator = new CardDeckGenerator();
+        Queue<Card> cards = cardDeckGenerator.generate();
 
         assertThat(cards.size()).isEqualTo(52);
     }
